requirements.txt
betterproto==1.2.5
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dialogflow==1.1.0
Flask==1.1.2
google-api-core==1.28.0
google-auth==1.30.1
googleapis-common-protos==1.53.0
grpcio==1.38.0
grpclib==0.4.1
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
multidict==5.1.0
nltk==3.6.2
numpy==1.20.3
packaging==20.9
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
scikit-learn==0.22.2.post1
scipy==1.6.3
six==1.16.0
stringcase==1.2.0
tqdm==4.61.0
urllib3==1.26.4
Werkzeug==1.0.1
wincertstore==0.2
